+++ .gitignore
@@ -3,3 +3,4 @@ bin/
 coverage/
 *.out
 *.test
+./ably-boomer
