+++ tasks/ably/sharded.go
@@ -1,4 +1,4 @@
-package tasks
+package ably
 
 import (
 	"context"
